--- src/pages/auth/SignUp.tsx
@@ -1,9 +1,0 @@
-const SignUp = () => {
-    return (
-        <div>
-            
-        </div>
-    );
-}
-
-export default SignUp;
